requirements.txt
fastapi>=0.135.0
uvicorn
python-dotenv
google-genai
google-cloud-aiplatform
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
requests
